实现/mcdc reload命令
